[kernels]: split moe_gemm_2stage monolith into a package
Structure-only split; generated code unchanged. The 3586-line
moe_gemm_2stage.py becomes a package with the three compile functions in
separate modules; public API is re-exported from __init__ so all importers
keep working:
  moe_gemm_2stage/{_gemm1,_gemm2,_reduction}.py

Function bodies are byte-identical to the pre-split module (verified), so the
trace-based DSL emits identical IR.
